@@ -1,4 +1,5 @@
-import { Component, Input, OnInit, } from '@angular/core';
+import { Component, OnInit, Input } from '@angular/core';
+
 import { Recipe } from '../../recipe.model';
 
 @Component({
@@ -10,9 +11,6 @@ export class RecipeItemComponent implements OnInit {
   @Input() recipe: Recipe;
   @Input() index: number;
 
-  constructor() { 
-  }
-
-  ngOnInit(): void {
+  ngOnInit() {
   }
 }
