@@ -1,6 +1,3 @@
 export class Ingredient {
-    constructor(
-        public name: string, 
-        public amount: number) { 
-    }
+  constructor(public name: string, public amount: number) {}
 }
