@@ -1,6 +1,12 @@
-import { Component, OnInit, OnDestroy, ViewChild } from '@angular/core';
+import {
+  Component,
+  OnInit,
+  OnDestroy,
+  ViewChild
+} from '@angular/core';
 import { NgForm } from '@angular/forms';
 import { Subscription } from 'rxjs';
+
 import { Ingredient } from '../../shared/ingredient.model';
 import { ShoppingListService } from '../shopping-list.service';
 
@@ -10,53 +16,53 @@ import { ShoppingListService } from '../shopping-list.service';
   styleUrls: ['./shopping-edit.component.css']
 })
 export class ShoppingEditComponent implements OnInit, OnDestroy {
-  @ViewChild('f', { static: false }) shoppingListForm: NgForm;
+  @ViewChild('f') slForm: NgForm;
   subscription: Subscription;
-  index: number;
-  editMode: boolean = false;
-  editIngredient: Ingredient;
+  editMode = false;
+  editedItemIndex: number;
+  editedItem: Ingredient;
 
-  constructor(private shoppingListService: ShoppingListService) { 
-  }
+  constructor(private slService: ShoppingListService) { }
 
-  ngOnDestroy(): void {
-    this.subscription.unsubscribe();
-  }
-
-  ngOnInit(): void {
-    this.subscription = this.shoppingListService.editingIngredientsEvent
-      .subscribe((index: number) => {
-        this.index = index;
-        this.editMode = true;
-        this.editIngredient = this.shoppingListService.getIngredientById(index);
-        this.shoppingListForm.setValue({
-          name: this.editIngredient.name,
-          amount: this.editIngredient.amount
-        })
-      });
-  }
-
-  onClear(): void {
-    this.editMode = false;
-    this.shoppingListForm.reset();
-  }
-
-  onDelete(): void {
-    this.shoppingListService.deleteIngredient(this.index);
-    this.onClear();
+  ngOnInit() {
+    this.subscription = this.slService.startedEditing
+      .subscribe(
+        (index: number) => {
+          this.editedItemIndex = index;
+          this.editMode = true;
+          this.editedItem = this.slService.getIngredient(index);
+          this.slForm.setValue({
+            name: this.editedItem.name,
+            amount: this.editedItem.amount
+          })
+        }
+      );
   }
 
   onSubmit(form: NgForm) {
-    const val = form.value;
-    const ingredient = new Ingredient(val.name, val.amount);
-
+    const value = form.value;
+    const newIngredient = new Ingredient(value.name, value.amount);
     if (this.editMode) {
-      this.shoppingListService.updateIngredient(this.index, ingredient);
+      this.slService.updateIngredient(this.editedItemIndex, newIngredient);
     } else {
-      this.shoppingListService.addIngredient(ingredient);
+      this.slService.addIngredient(newIngredient);
     }
-
     this.editMode = false;
     form.reset();
   }
+
+  onClear() {
+    this.slForm.reset();
+    this.editMode = false;
+  }
+
+  onDelete() {
+    this.slService.deleteIngredient(this.editedItemIndex);
+    this.onClear();
+  }
+
+  ngOnDestroy() {
+    this.subscription.unsubscribe();
+  }
+
 }
