@@ -1,5 +1,6 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Subscription } from 'rxjs';
+
 import { Ingredient } from '../shared/ingredient.model';
 import { ShoppingListService } from './shopping-list.service';
 
@@ -9,25 +10,26 @@ import { ShoppingListService } from './shopping-list.service';
   styleUrls: ['./shopping-list.component.css']
 })
 export class ShoppingListComponent implements OnInit, OnDestroy {
-  private ingredientsSubscription: Subscription;
   ingredients: Ingredient[];
+  private subscription: Subscription;
 
-  constructor(private shoppingListService : ShoppingListService) { 
+  constructor(private slService: ShoppingListService) { }
+
+  ngOnInit() {
+    this.ingredients = this.slService.getIngredients();
+    this.subscription = this.slService.ingredientsChanged
+      .subscribe(
+        (ingredients: Ingredient[]) => {
+          this.ingredients = ingredients;
+        }
+      );
   }
 
   onEditItem(index: number) {
-    this.shoppingListService.editingIngredientsEvent.next(index);
+    this.slService.startedEditing.next(index);
   }
 
-  ngOnDestroy(): void {
-    this.ingredientsSubscription.unsubscribe();
-  }
-
-  ngOnInit(): void { 
-    this.ingredients = this.shoppingListService.getIngredients();
-    this.ingredientsSubscription = this.shoppingListService.ingredientsChangedEvent.subscribe(
-      (ingredients: Ingredient[]) => {
-        this.ingredients = ingredients;
-      })
+  ngOnDestroy() {
+    this.subscription.unsubscribe();
   }
 }
